refactor(dashboard): extract test tile fixtures to shared file
